+++ package.json
@@ -15,9 +15,14 @@
         "rollup-plugin-node-externals": "^7.1.3",
         "typescript": "^5.6.3",
         "vite": "^5.4.10",
+        "vite-plugin-inspect": "^0.8.7",
         "vitest": "^2.1.4"
     },
     "peerDependencies": {
         "lodash-es": "*"
+    },
+    "dependencies": {
+        "@types/lodash-es": "^4.17.12",
+        "unplugin": "^1.15.0"
     }
 }
