+++ src/types.ts
@@ -1,0 +1,4 @@
+export interface Options {
+    forceIncludeFns?: string[];
+    excludes?: RegExp[];
+}
