+++ vite.config.ts
@@ -1,18 +1,20 @@
 import { defineConfig } from "vite";
-import { LodashToEs } from "./src";
+import vitePlugin from "./src/vite";
 import { nodeExternals } from "rollup-plugin-node-externals";
+import Inspect from "vite-plugin-inspect";
 export default defineConfig({
     plugins: [
-        // nodeExternals({
-        //     devDeps: true,
-        // }),
-        LodashToEs(),
+        nodeExternals({
+            devDeps: true,
+        }),
+        vitePlugin(),
+        Inspect({ build: true,}),
     ],
     build: {
         target: "esnext",
         minify: false,
         lib: {
-            entry: "./test/example.js",
+            entry: ["./src/index.ts","./src/vite.ts"],
             name: "lodash-to-es",
             fileName: "lodash-to-es",
             formats: ["es"],
